@holiman 2020-06-10
This is now being benchmarked: https://geth-bench.ethdevops.io/d/Jpk-Be5Wk/dual-geth?orgId=1&from=1591781494201&to=now&refresh=5s&var-exp=bench01&var-master=bench02&var-percentile=50 . 
`bench03` is this PR, `bench04` is master.